refactor: 패키지 이름 수정
의존성 문제로 인해 실행되지 않는 문제가 있었음
패키지 이름을 통일하고 의존성을 제거하여 실행 가능하도록 문제를 해결함
