+++ server/src/main/java/com/genesisairport/reservation/controller/ReservationController.java
@@ -1,7 +1,7 @@
 package com.genesisairport.reservation.controller;
 
-import com.genesisairport.reservation.Response.ReservationListAbstract;
-import com.genesisairport.reservation.Response.ReservationPostResponse;
+import com.genesisairport.reservation.response.ReservationListAbstract;
+import com.genesisairport.reservation.response.ReservationPostResponse;
 
 import com.genesisairport.reservation.common.DataResponseDto;
 import com.genesisairport.reservation.response.ReservationResponse;
